+++ src/hkim0331/util/core.clj
@@ -220,5 +220,17 @@
   (reverse-all [1 [[2 3] 4 [5 [[6]]] 7] 8 9])
   :rcf)
 
-(defn abbrev [s]
-  ())
+(defn binary-search [v value]
+  (loop [low 0 high (dec (count v))
+         depth 0]
+    (if (<= high (inc low))
+      (cond (= value (v low)) low
+            (= value (v high)) high
+            :else nil)
+      (let [middle (quot (+ low high) 2)]
+        (if (< (v middle) value)
+          (recur (inc middle) high (inc depth))
+          (recur low middle (inc depth)))))))
+
+(comment
+  (binary-search (vec (range 1000)) 500))